Update API client to use generic method for fetching cats and images
Updated ApiClient to use a generic method GetAsync<T> for fetching cats and images.
GetCatsAsync and GetCatImageAsync are now implemented using the generic method.
GetCatsAsync now returns a list of CatApiResponse objects.
GetCatImageAsync now returns a byte array of the image data.
GetAsync<T> is used to fetch cats and images from the API.
The GetByteArrayAsync method is used to retrieve the image data.
GetCatsAsync and GetCatImageAsync are now more efficient and maintainable.